+++ elec631/20171116_0107_duck_shooting/src/Matrix.cpp
@@ -1,12 +1,12 @@
 #include <iomanip>  //For std::set(w)
 #include "Matrix.hpp"
 //Constructor
-Matrix_NS::Matrix::Matrix(const int r, const int c) : row_(r), column_(c), defaultVal_(0){
+Matrix_NS::Matrix::Matrix(const uint64_t r, const uint64_t c) : row_(r), column_(c), defaultVal_(0){
     //std::cout << "\nConstructor1 is called...\n\n";
     createMatrix(r, c, 0);
 }
 //Constructor
-Matrix_NS::Matrix::Matrix(const int r, const int c, const double val) : row_(r), column_(c), defaultVal_(val) {
+Matrix_NS::Matrix::Matrix(const uint64_t r, const uint64_t c, const double val) : row_(r), column_(c), defaultVal_(val) {
     //std::cout << "\nConstructor2 is called...\n\n";
     createMatrix(r, c, val);
 }
@@ -19,8 +19,8 @@ Matrix_NS::Matrix::~Matrix(){
 Matrix_NS::Matrix::Matrix(const Matrix& other) : row_(other.row_), column_(other.column_), defaultVal_(other.defaultVal_){
     //std::cout << "\nCopy Constructor is called...\n\n";
     createMatrix(other.row_, other.column_, other.defaultVal_);
-    for(int i =0; i<other.row_; i++){
-        for (int j=0; j<other.column_; j++){
+    for(uint64_t i =0; i<other.row_; i++){
+        for (uint64_t j=0; j<other.column_; j++){
             this->matrix_.at(i).at(j) = other.matrix_.at(i).at(j);
         }
     }
@@ -35,8 +35,8 @@ Matrix_NS::Matrix& Matrix_NS::Matrix::operator=( Matrix& rhs){
         this->column_ = rhs.column_;
         this->defaultVal_ = rhs.defaultVal_;
 
-        for(int i=0; i<rhs.row_; i++){
-            for (int j=0; j<rhs.column_; j++){
+        for(uint64_t i=0; i<rhs.row_; i++){
+            for (uint64_t j=0; j<rhs.column_; j++){
                 this->matrix_.at(i).at(j) = rhs.matrix_.at(i).at(j);
             }
         }
@@ -45,28 +45,28 @@ Matrix_NS::Matrix& Matrix_NS::Matrix::operator=( Matrix& rhs){
 }
 
 //Getters and setters
-const int Matrix_NS::Matrix::R_() const {      return row_;                }
-const int Matrix_NS::Matrix::C_() const {      return column_;             }
-const int Matrix_NS::Matrix::numel() const {   return row_*column_;        }
+const uint64_t Matrix_NS::Matrix::R_() const {      return row_;                }
+const uint64_t Matrix_NS::Matrix::C_() const {      return column_;             }
+const uint64_t Matrix_NS::Matrix::numel() const {   return row_*column_;        }
 const std::vector< std::vector<double> >& Matrix_NS::Matrix::M_() const {    return matrix_; }
-Matrix_NS::Matrix& Matrix_NS::Matrix::M_(int row, int col, double val){
+Matrix_NS::Matrix& Matrix_NS::Matrix::M_(uint64_t row, uint64_t col, double val){
     matrix_.at(row).at(col) = val;
     return *this;
 }
 
 //Member operator overloading for column major matrix indexing
-double Matrix_NS::Matrix::operator()(const int index){
+double Matrix_NS::Matrix::operator()(const uint64_t index){
     return matrix_.at( index%row_ ).at( index/row_ );
 }
 
-void Matrix_NS::Matrix::operator()(const int index, const double value){
+void Matrix_NS::Matrix::operator()(const uint64_t index, const double value){
     matrix_.at( index%row_ ).at( index/row_ ) = value;
 }
 //Member operator overloading for basic mathematic matrix operations
 Matrix_NS::Matrix& Matrix_NS::Matrix::operator+(const Matrix& rhs){
     //std::cout << "Inside member + overloading...\n";
-    for(int i=0; i<rhs.row_; i++){
-        for(int j=0; j<rhs.column_; j++){
+    for(uint64_t i=0; i<rhs.row_; i++){
+        for(uint64_t j=0; j<rhs.column_; j++){
             this->matrix_.at(i).at(j) = this->matrix_.at(i).at(j) + rhs.matrix_.at(i).at(j);
         }
     }
@@ -76,8 +76,8 @@ Matrix_NS::Matrix& Matrix_NS::Matrix::operator+(const Matrix& rhs){
 //Matrix_NS::Matrix& Matrix_NS::Matrix::operator*(const Matrix& rhs){
 Matrix_NS::Matrix& Matrix_NS::Matrix::operator*(double scalar){
     //std::cout << "Inside member * overloading...\n";
-    for(int i=0; i<this->row_; i++){
-        for(int j=0; j<this->column_; j++){
+    for(uint64_t i=0; i<this->row_; i++){
+        for(uint64_t j=0; j<this->column_; j++){
             this->matrix_.at(i).at(j) = this->matrix_.at(i).at(j) * scalar;
         }
     }
@@ -87,11 +87,11 @@ Matrix_NS::Matrix& Matrix_NS::Matrix::operator*(double scalar){
 //Matrix_NS::Matrix& Matrix_NS::Matrix::operator/(const Matrix& rhs){}
 
 //Create matrix with desired values
-const void Matrix_NS::Matrix::createMatrix(const int r, const int c, double val){
+const void Matrix_NS::Matrix::createMatrix(const uint64_t r, const uint64_t c, double val){
     this->matrix_.resize(r);
-    for(int i=0; i<r; i++){
+    for(uint64_t i=0; i<r; i++){
         this->matrix_.at(i).resize(c);
-        for(int j=0; j<c; j++){
+        for(uint64_t j=0; j<c; j++){
             this->matrix_.at(i).at(j) = val;
         }
     }
@@ -99,8 +99,8 @@ const void Matrix_NS::Matrix::createMatrix(const int r, const int c, double val)
 //Print matrix for debugging issues
 const void Matrix_NS::Matrix::printMatrix() const{
     std::cout << "Printing matrix...\n";
-    for(int i=0; i<row_; i++){
-        for(int j=0; j<column_; j++){
+    for(uint64_t i=0; i<row_; i++){
+        for(uint64_t j=0; j<column_; j++){
             //std::cout  << matrix_.at(i).at(j) << " ";
             //std::cout <<std::fixed<< matrix_.at(i).at(j) << " ";
             std::cout << std::fixed << std::setw(5) << std::setprecision(3) << matrix_.at(i).at(j) << " ";
@@ -114,8 +114,8 @@ Matrix_NS::Matrix& Matrix_NS::operator+(double scalar, const Matrix& rhs){
 //    std::cout << "\nInside nonmember + overloading...\n";
     Matrix temp(rhs);
 //    temp.printMatrix();
-    for(int i=0; i<temp.R_(); i++){
-        for(int j=0; j<temp.C_(); j++){
+    for(uint64_t i=0; i<temp.R_(); i++){
+        for(uint64_t j=0; j<temp.C_(); j++){
             temp.M_(i, j, temp.M_().at(i).at(j)+scalar);
         }
     }
@@ -128,8 +128,8 @@ Matrix_NS::Matrix& Matrix_NS::operator*(double scalar, const Matrix& rhs){
 //    std::cout << "Inside nonmember left * overloading...\n";
     Matrix temp(rhs);
 //    temp.printMatrix();
-    for(int i=0; i<temp.R_(); i++){
-        for(int j= 0; j< temp.C_(); j++){
+    for(uint64_t i=0; i<temp.R_(); i++){
+        for(uint64_t j= 0; j< temp.C_(); j++){
             temp.M_(i, j, temp.M_().at(i).at(j)*scalar);
         }
     }
@@ -143,8 +143,8 @@ Matrix_NS::Matrix& Matrix_NS::operator*(const Matrix& rhs, double scalar){
 //    std::cout << "Inside nonmember right * overloading...\n";
     Matrix temp(rhs);
 //    temp.printMatrix();
-    for(int i=0; i<rhs.R_(); i++){
-        for(int j= 0; j< rhs.C_(); j++){
+    for(uint64_t i=0; i<rhs.R_(); i++){
+        for(uint64_t j= 0; j< rhs.C_(); j++){
 //            double d = rhs.M_().at(i).at(j);
 //            rhs.M_(i, j, d*scalar);
             temp.M_(i, j, temp.M_().at(i).at(j)*scalar);
